+++ Source/ACE.Database/WorldDatabase.cs
@@ -19,7 +19,8 @@ namespace ACE.Database
             GetAnimationOverridesByObject,
             GetTextureOverridesByObject,
             GetCreatureDataByWeenie,
-            InsertCreatureStaticLocation
+            InsertCreatureStaticLocation,
+            GetPortalDestination
         }
 
         protected override Type PreparedStatementType => typeof(WorldPreparedStatement);
@@ -28,6 +29,7 @@ namespace ACE.Database
         {
             AddPreparedStatement(WorldPreparedStatement.TeleportLocationSelect, "SELECT `location`, `cell`, `x`, `y`, `z`, `qx`, `qy`, `qz`, `qw` FROM `teleport_location`;");
             // ConstructStatement(WorldPreparedStatement.GetWeenieClass, typeof(BaseAceObject), ConstructedStatementType.Get);
+			ConstructStatement(WorldPreparedStatement.GetPortalDestination, typeof(PortalDestination), ConstructedStatementType.Get);
             ConstructStatement(WorldPreparedStatement.GetObjectsByLandblock, typeof(AceObject), ConstructedStatementType.GetList);
             ConstructStatement(WorldPreparedStatement.GetCreaturesByLandblock, typeof(AceCreatureStaticLocation), ConstructedStatementType.GetList);
             ConstructStatement(WorldPreparedStatement.GetWeeniePalettes, typeof(WeeniePaletteOverride), ConstructedStatementType.GetList);
@@ -56,6 +58,21 @@ namespace ACE.Database
             }
 
             return locations;
+        }
+
+        public PortalDestination GetPortalDestination(uint weenieClassId)
+        {
+            PortalDestination portalDestination = new PortalDestination();
+            Dictionary<string, object> criteria = new Dictionary<string, object>();
+            criteria.Add("weenieClassId", weenieClassId);
+            if (ExecuteConstructedGetStatement(WorldPreparedStatement.GetPortalDestination, typeof(PortalDestination), criteria, portalDestination))
+            {
+                return portalDestination;
+            }
+            else
+            {
+                return null;
+            }
         }
 
         public List<AceObject> GetObjectsByLandblock(ushort landblock)